fix: problems with SSR in fullwidth-styled page
