+++ apps/demo-sidebar/content/docs/root.mdx
@@ -1,6 +1,8 @@
 ---
 title: 'Root'
 description: 'SEO description'
+fullWidth: true
+disableBreadcrumbs: true
 ---
 
 Perex
