+++ apps/demo-sidebar/src/pages/index.js
@@ -3,7 +3,7 @@ import { Page, Seo } from 'gatsby-theme-fast-ai-sidebar';
 import { Heading } from '@fast-ai/ui-components';
 
 const Index = props => (
-	<Page {...props}>
+	<Page fullWidth {...props}>
 		<Seo title="Home" />
 
 		<Heading>Welcome</Heading>
