+++ packages/gatsby-theme-fast-ai-sidebar/src/intl.js
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { IntlContextConsumer } from 'gatsby-plugin-intl';
+
+export const useIntlContext = () => useContext(IntlContextConsumer._context);
