+++ packages/gatsby-theme-fast-ai-sidebar/src/templates/MdxPage.js
@@ -65,6 +65,7 @@ export const pageQuery = graphql`
 				description
 				tableOfContentsDepth
 				disableTableOfContents
+				disableBreadcrumbs
 				fullWidth
 			}
 			id
